fix intermittent reconfig issue with consumer
