--- dev/com.ibm.ws.security.jwt_fat.consumer/publish/servers/com.ibm.ws.security.jwt_fat.consumer/configs/server_minimumConfig.xml
@@ -1,15 +1,0 @@
-
-<server>
-
-	<featureManager>
-		<feature>timedexit-1.0</feature>
-		<feature>servlet-3.1</feature>
-		<feature>jwt-1.0</feature>
-		<feature>jsonp-1.0</feature>
-	</featureManager>
-
-	<include location="${shared.config.dir}/fatTestPorts.xml" />
-
-	<javaPermission className="java.util.PropertyPermission" actions="read" name="*"/>
-
-</server>
